Renamed and restructured files in ws package
